requirements.txt
fastapi>=0.110.0
uvicorn>=0.30.0
pydantic>=2.10.0
huggingface_hub>=1.7.0
pillow>=10.3.0
pytest>=8.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
python-multipart>=0.0.12